Add download test script
Add python script to downoad files from Permament.
